Changed main dump loop to be more similar to a timestep controlled loop
A time t, delta dt and end time maxT are used to control the main loop.
If no compute is being used, the loop will revert to the behaviour seen previously where we just did consecutive dumps.
If there is compute used, the compute route will work for the specified time and then increase the current time by dt.