Added some MACSIO_MAIN_* vars to tstlog and tsttiming to fix build problem
MACSIO_MAIN_Rank,Size,Comm are used in macsio_work functions, which are linked into the tests because of the way the makefile grabs all of the macsio object files (except from macsio_main). Linker was failing because these vars weren't declared in the linked files so added them to the top of tsttiming and tstlog so their build doesn't fail. Needs a better solution ideally.