Added fix to the partitioning
The choose_part_count function that is important for creating an irregular mesh partitioning wasn't behaving as expected. Due to operator precedence the values of R and Q weren't being decremented, instead the operation was being applied to the pointer itself and hence odd things were happening. Fixed with some simple bracketing.